- add connect-laser script to ubuntu-16.04 dir  - add connect-rosaroa script to ubuntu-16-04 dir  - install csm from apt-get (change install-csm script)

+++ ubuntu-16-04/install-csm.sh
@@ -7,19 +7,26 @@ FOLDER_NAME=csm-$LIBCSM_VER
 FILE_NAME=$FOLDER_NAME.zip
 LINK=https://github.com/AndreaCensi/csm/archive/$LIBCSM_VER.zip
 CMAKE_FLAGS="-DCMAKE_INSTALL_PREFIX=$PREFIX -DCMAKE_BUILD_TYPE=Release"
+FROM_GIT=false
 
 
 sudo apt-get install libgsl-dev -y
 
-sudo rm -rf $PROJECT_DIR/$FOLDER_NAME    
-cd ~/Downloads
-wget -O $FILE_NAME $LINK
-mkdir -p $PROJECT_DIR
-unzip $FILE_NAME -d $PROJECT_DIR
-rm -f ~/Downloads/$FILE_NAME
-cd $PROJECT_DIR/$FOLDER_NAME
-mkdir build && cd build
-cmake $CMAKE_FLAGS ..
-make -j7       
-sudo make install -j7
+if [ "$FROM_GIT" = True ]; then
+    sudo rm -rf $PROJECT_DIR/$FOLDER_NAME    
+    cd ~/Downloads
+    wget -O $FILE_NAME $LINK
+    mkdir -p $PROJECT_DIR
+    unzip $FILE_NAME -d $PROJECT_DIR
+    rm -f ~/Downloads/$FILE_NAME
+    cd $PROJECT_DIR/$FOLDER_NAME
+    mkdir build && cd build
+    cmake $CMAKE_FLAGS ..
+    make -j7       
+    sudo make install -j7
+else
+    sudo apt-get install ros-kinetic-csm
+fi
+
+
